Prevent data loss if header begins with a number
If a header's text begins with a number, this script will remove the number when run. For example, if a header starts with a date such as "4/5/2020", it will become "1. /5/2020" when the script is run. This commit modifies the regex that is used to extract the numbering from the header so that this does not happen.